refactor: moved homepage from /app to /components
