--- src/app/(homepage)/SectionArrivals/index.ts
@@ -1,4 +1,0 @@
-export { default as NewArrivals } from "@/app/(homepage)/SectionArrivals/constants/NewArrivals";
-export { default as SectionArrivals } from "@/app/(homepage)/SectionArrivals/components/SectionArrivals";
-
-export type { NewArrival } from "@/app/(homepage)/SectionArrivals/types";
