--- src/app/(homepage)/SectionArrivals/types/index.ts
@@ -1,5 +1,0 @@
-export type NewArrival = {
-  src: string;
-  name: string;
-  href: string;
-};
