--- src/app/(homepage)/SectionDiscount/index.ts
@@ -1,7 +1,0 @@
-export { default as SectionDiscount } from "@/app/(homepage)/SectionDiscount/components/SectionDiscount";
-export {
-  DiscountedBottom,
-  DiscountedUpper,
-} from "@/app/(homepage)/SectionDiscount/constants/DiscountedCategories";
-
-export type { DiscountedItem } from "@/app/(homepage)/SectionDiscount/types";
