--- src/app/(homepage)/SectionDiscount/types/index.ts
@@ -1,15 +1,0 @@
-export type DiscountedItem = {
-  image: {
-    src: string;
-    className?: string;
-  };
-
-  content: {
-    title: string;
-    subtitle: string;
-    discount: number;
-    className?: string;
-  };
-
-  isLimited?: boolean;
-};
