--- src/app/(homepage)/SectionFeatured/index.ts
@@ -1,4 +1,0 @@
-export { default as SectionFeatured } from "@/app/(homepage)/SectionFeatured/components/SectionFeatured";
-export { default as FeaturedCategories } from "@/app/(homepage)/SectionFeatured/constants/FeaturedCategories";
-
-export type { FeaturedCategory } from "@/app/(homepage)/SectionFeatured/types";
