--- src/app/(homepage)/SectionFeatured/types/index.ts
@@ -1,13 +1,0 @@
-export type FeaturedCategory = {
-  image: {
-    src: string;
-    className?: string;
-  };
-
-  content: {
-    href: string;
-    title: string;
-    subtitle: string;
-    discount: number;
-  };
-};
