--- src/app/(homepage)/SectionTestimonials/index.ts
@@ -1,5 +1,0 @@
-export { default as Testimonials } from "@/app/(homepage)/SectionTestimonials/constants/Testimonials";
-
-export { default as SectionTestimonials } from "@/app/(homepage)/SectionTestimonials/components/SectionTestimonials";
-
-export type { Testimonial } from "@/app/(homepage)/SectionTestimonials/types";
