--- src/app/(homepage)/Slides/index.ts
@@ -1,1 +1,0 @@
-export { default as HomeSlides } from "@/app/(homepage)/Slides/constants/HomeSlides";
