+++ src/components/Homepage/components/SectionDiscount.tsx
@@ -2,10 +2,7 @@ import cn from "@/utils/cn";
 import { Section } from "@/components/Section";
 import { ArrowIcon } from "@/components/Icons";
 import { AbsoluteCard } from "@/components/AbsoluteCard";
-import {
-  DiscountedBottom,
-  DiscountedUpper,
-} from "@/app/(homepage)/SectionDiscount";
+import { DiscountedBottom, DiscountedUpper } from "@/components/Homepage";
 
 export default function SectionDiscount() {
   return (
