+++ src/components/Homepage/index.ts
@@ -1,0 +1,19 @@
+export { default as SectionArrivals } from "@/components/Homepage/components/SectionArrivals";
+export { default as SectionDiscount } from "@/components/Homepage/components/SectionDiscount";
+export { default as SectionFeatured } from "@/components/Homepage/components/SectionFeatured";
+export { default as SectionTestimonials } from "@/components/Homepage/components/SectionTestimonials";
+
+export {
+  DiscountedBottom,
+  DiscountedUpper,
+} from "@/components/Homepage/constants/DiscountedCategories";
+export { default as FeaturedCategories } from "@/components/Homepage/constants/FeaturedCategories";
+export { default as HomeSlides } from "@/components/Homepage/constants/HomeSlides";
+export { default as NewArrivals } from "@/components/Homepage/constants/NewArrivals";
+export { default as Testimonials } from "@/components/Homepage/constants/Testimonials";
+
+export type {
+  DiscountedItem,
+  FeaturedCategory,
+  NewArrival,
+} from "@/components/Homepage/types";
