+++ src/components/Testimonial/index.tsx
@@ -1,1 +1,3 @@
-export { default as Testimonial } from "@/components/Testimonial/Testimonial";
+export { default as Testimonial } from "@/components/Testimonial/components/Testimonial";
+
+export type { Testimonial as TTestimonial } from "@/components/Testimonial/types";
